+++ HISTORY.md
@@ -13,6 +13,7 @@
 * extensions: tweaked variable declaration extension primitive, commented out palette refresh prim
 * tweaked make-variables library
 * tweaked strings library
+* extensions: added color library dropdown menu
 
 ### 2021-06-17
 * extensions: added APL extension primitives